Add files via upload
Make a class named sweet with a data member to calculate the number of sweets in a
basket. Create two other classes named Chocolate and Ice cream to calculate the
number of Chocolate and Ice cream in the basket. Print the number of sweets of each
type and the total number of sweets in the basket.